feat: Update to 0.9.2 (#43)

+++ src/network/receipt_envelope.rs
@@ -47,7 +47,7 @@ where
         }
     }
 
-    fn cumulative_gas_used(&self) -> u128 {
+    fn cumulative_gas_used(&self) -> u64 {
         match self {
             ReceiptEnvelope::Native(re) => re.cumulative_gas_used(),
             ReceiptEnvelope::Eip712(re) => re.cumulative_gas_used(),
